Replace sqrt( by sqrtf(.

+++ flight/PiOS/Common/pios_mpxv5004.c
@@ -88,7 +88,7 @@ float PIOS_MPXV5004_ReadAirspeed(uint8_t airspeedADCPin)
 	}
 	
 	//Compute calibraterd airspeed, as per http://en.wikipedia.org/wiki/Calibrated_airspeed
-	float calibratedAirspeed = A0*sqrt(5.0f*(pow(Qc/P0+1.0f,POWER)-1.0f));
+	float calibratedAirspeed = A0*sqrtf(5.0f*(pow(Qc/P0+1.0f,POWER)-1.0f));
 	
 	//Upper bound airspeed. No need to lower bound it, that comes from Qc
 	if (calibratedAirspeed > 80) { //in [m/s]
